--- Assets/FreeAssets/FastMesh/Scripts/SceneViewText.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9549876df73b7e04bac73492fd5b95c4
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
