+++ Assets/IMPORTED/Plugins/FMOD/platforms/win/lib/x86_64/fmodstudioL.dll.meta
@@ -22,11 +22,12 @@ PluginImporter:
       enabled: 1
       settings:
         CPU: x86_64
+        DefaultValueInitialized: true
         OS: Windows
   - first:
       Standalone: Win64
     second:
-      enabled: 1
+      enabled: 0
       settings: {}
   userData: 
   assetBundleName:
