+++ Assets/Scripts/DevDebug.cs
@@ -24,6 +24,11 @@ public class DevDebug : MonoBehaviour
             SceneManager.LoadScene(SceneManager.GetActiveScene().buildIndex);
         }
 
+        if (Input.GetKeyDown(KeyCode.Escape))
+        {
+            Application.Quit();
+        }
+
     }
 
 }
